amd64 port

+++ pps/src/lieferscheinliste.cc
@@ -166,7 +166,7 @@ void lieferscheinliste::fill_tree()
      for (std::vector<LieferscheinEntryBase>::const_iterator j=LL.begin((*i)->Id());
      		j!=LL.end((*i)->Id()); ++j)
         datavec.push_back(new Data_LListe(*i,LieferscheinEntry(*j),R,
-             	KumVal(reinterpret_cast<int>(date_cumulate->get_menu()->get_active()->get_user_data()))
+             	KumVal(reinterpret_cast<long>(date_cumulate->get_menu()->get_active()->get_user_data()))
              	));
       progressbar->set_percentage(count/size);
       while(Gtk::Main::events_pending()) Gtk::Main::iteration() ;
